# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Empleo",
        "description": "Sistema de gestión de ofertas de empleo y seguimiento de candidatos.",
        "contact": {
            "email": "admin@centro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/api/perfil": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Obtener perfil del usuario",
                "operationId": "getPerfilDetallado",
                "responses": {
                    "200": {
                        "description": "Éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil cargado"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/perfil/direccion": {
            "post": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Guardar o actualizar dirección",
                "operationId": "d105ef4335bd79bf94a6e976120ce39b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "linea1",
                                    "ciudad",
                                    "provincia",
                                    "codigoPostal",
                                    "visible"
                                ],
                                "properties": {
                                    "linea1": {
                                        "type": "string"
                                    },
                                    "linea2": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ciudad": {
                                        "type": "string"
                                    },
                                    "provincia": {
                                        "type": "string"
                                    },
                                    "codigoPostal": {
                                        "type": "string"
                                    },
                                    "visible": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "200": {
                        "description": "Actualizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/perfil/editar": {
            "patch": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Actualizar datos del perfil",
                "operationId": "d748d8cacbe3cc92233fa0e30b1fa57a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cif": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "localidad": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "experienciaLaboral": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "situacion": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actualizado correctamente"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/perfil/situaciones": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Listar situaciones",
                "operationId": "f15fe603fd120f22058a66e9485fe70a",
                "responses": {
                    "200": {
                        "description": "Lista de situaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "Authorization",
                "in": "header",
                "example": "Bearer 19|tBfCzyfhD1Ib3VA5HPyHfR6CjrIoyT5VPq7s0mMnb19cdfbe"
            }
        }
    },
...
}